GoogleDinoBot/requirements.txt
numpy==1.16.1
Pillow==5.4.1
PyAutoGUI==0.9.41

